+++ src/layouts/page.njk
@@ -72,6 +72,6 @@
     <div class="page__content{% if isIndex %} page__content--inner{% endif %}">
         {{ content | safe }}
     </div>
-    <script src="/scripts/scripts.js"></script>
+    <script type="module" src="/scripts/scripts.js"></script>
 </body>
 </html>
