+++ src/styles/styles.css
@@ -25,6 +25,7 @@
 @import 'blocks/article-card.css';
 @import 'blocks/conference-link.css';
 @import 'blocks/video.css';
+@import 'blocks/tooltip.css';
 
 /* Globals */
